requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
# scikit-learn
transformers
requests
# tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
# tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
# tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
torch

# image processing
Python-IO
Pillow

# API
gcsfs
# google-cloud-storage
# mlflow
fastapi
uvicorn
python-multipart


# utilities
six>=1.14
joblib
memoized-property
termcolor

# user-interface
streamlit_option_menu
gTTS
#SpeechRecognition
#PyAudio
# time
# io
